[vm] Consistently preserve constant representations when inlining.
Previously, when the inliner copied constants from caller to callee
graphs when creating parameter stubs, it always created a boxed version
of the constant, even if the constant is unboxed in the caller graph.
However, during the final steps of inlining, it would then use the
representation of the constant when copying all used constants from the
callee graph back to the caller graph.

This CL makes this behavior consistent so that the original
representation is also maintained when copying constants from the
caller graph to the callee graph.

This CL adds canonicalization for BoxInteger that replaces
BoxInteger<from>(UnboxedConstant<to>(i)) with Constant(i) if to is an
unboxed integer representation and i is representable in from.
(Previously, this was done for BoxInt64, but not for other BoxInteger
subclasses.)

TEST=vm/cc/IL_Canonicalize_BoxIntegerUnboxedConstant

Change-Id: I992098c0d47c4185792d983749fe6cf74eeef106
Cq-Include-Trybots: luci.dart.try:vm-aot-linux-release-simarm_x64-try,vm-aot-linux-debug-x64-try,vm-aot-mac-release-arm64-try,vm-aot-mac-product-arm64-try,vm-aot-linux-product-x64-try,vm-aot-dwarf-linux-product-x64-try,vm-aot-linux-debug-simarm_x64-try,vm-linux-debug-x64-try,vm-mac-debug-arm64-try,vm-mac-release-arm64-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/321980
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Commit-Queue: Tess Strickland <sstrickl@google.com>