[analysis_server] Fix == overrides being truncated in LSP
The LSP completion mapping currently does some manipulation of values to fit better into the different fields LSP supports (to improve formatting and filtering). The code that splits the label on `=>` was also matching `==` in overrides but should not.

This is a simple fix to the regex to solve the issue, although longer-term (or at least if we see other issues as a result of this manipulation) this logic should be pushed up to the source so we have reliable fields and don't need to split things this way.

Fixes https://github.com/dart-lang/sdk/issues/56336

Change-Id: I056489df5aba417cab9162e027eb97579f71756b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378200
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Phil Quitslund <pquitslund@google.com>