[cfe] Remove ForElement from supertypes of PatternForElement
This CL makes the handling of the two elements to be free from a
specialized discipline where the implementor needed to remember that
PatternForElement should be handled every time ForElement is handled.

A test case is added for something that was fixed in one of the other
recent CLs.

Change-Id: I612667055d4036aa301060265ed63546067fcdb9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378143
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>