Do simple type tests using T in non-external methods in js_interop_non_external_test
It's possible with changes in how we handle generics in
DDC that T might not be the type we expect at runtime. So,
we should test to make sure that it is by doing some simple
type tests in the body of non-external methods of JS interop
classes.

Change-Id: I0220b2bced4e5769e3728833204dfb3ac19f26d3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378572
Reviewed-by: Nicholas Shahan <nshahan@google.com>
Commit-Queue: Srujan Gaddam <srujzs@google.com>