Parts. Enhance FileSystemState_PubPackageTest for parts, do updates.
I wanted to check that the API signature of the class depends on
the imports/exports in parts.

And rename the tests methods, not sure how to have the most
consistent naming here. This is not 'newFile' anymore, because we
also refresh, but most other tests are 'newFile'. Sigh...

Change-Id: Ibc692357a8d884994e1f33f0b2140dfffb40767e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378578
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>