Adjust omit_obvious_local_variable_types
This CL adjusts the lint to ensure independence of inference for every
expression whose type is considered obvious. This is useful because it
means that we will never ask anyone to remove a type annotation such
that the resulting declaration will have a different type.

The CL also makes a cast expression (`e as T`) obvious.

Change-Id: Idd2265ac1d4f2ad4491b5a61938a9b125ade3a47
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378941
Commit-Queue: Erik Ernst <eernst@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>