[dart2wasm] Introduce `@pragma('wasm:static-dispatch')` and enable inlining of polymorphic specializations
Certain indirect calls dart2wasm is unable to devirtualize. Some of them
may target callees that are really small, where the overhead of an
indirect call is larger than the body of the function.

=> We add a `@pragma('wasm:static-dispatch')` which can be used to
annotate such methods

=> Code generator will then call a polymorphic specializer helper
function which switches on class-id and makes direct calls to the
targets.

=> Add inlining ability to the polymorphic helper functions
=> Make us inline them if there's <= 2 polymorphic targets.

=> Make use of this in `_Type.checkInstance()` where we want 2 specific
implementations be dispatched to statically the rest via a indirect call
fallback.

Change-Id: Ifaf517e1a82540672fa509e71b7cfa465d925237
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/378940
Reviewed-by: Ömer Ağacan <omersa@google.com>