[dart2wasm] Clean-up jsifyRaw checks
- It's common to jsify numbers and lists, so those cases
should be moved up as far as they can.
- It's uncommon for the wrapped function check to be used
as it's only relevant to allowInterop, which only a handful
of packages can even use.
- JS typed array wrappers can be grouped together under one
type-check. Note that even though the conversion is the same
for all the wrapper types, we still need to check the
individual types as there are subtypes we don't want to handle.
We do a similar grouping for the non-wrapper TypedData types.
- Clean up the if-else logic so it's a bit more concise.

Change-Id: Ibfcd135a6ac5e2edbaac263d03ea6832c4131c1d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/379143
Reviewed-by: Martin Kustermann <kustermann@google.com>