[dart2wasm] Fix multi root invocations
`dart2wasm` supports the `--multi-root` argument and respective URIs as
inputs.
`build_runner` uses the multi root argument when compiling with dart2js
as it allows a file system layout like this, placing `lib/` contents of
packages into folders next to non-lib assets of the root package being
compiled:

```
input/
  args/
  async/
  .../
  other_packages/
  web/
    index.html
    main.dart
```

While it looks like `dart2wasm` is supposed to support the same
arguments as well, the way it parses URI flags and a pre-check in
`dartdev` unfortunately prevent this.
This fixes `dartdev` to not check for inputs when a multi-root argument
is given, as the paths are not physical paths from the operating system
in that case.
It also fixes the `UriOption` in `dart2wasm` to be consistent with how
input arguments are resolved, allowing multi-root scheme URIs to be
used as arguments for e.g. `--packages`.

Change-Id: If751a5ee58938d5c7c62e27a69957456d8eea860
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377662
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Kevin Moore <kevmoo@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>