[test_runner] Add option to run browser tests with ddc module format
With hot reload, a new module format is being introduced that's
repurposing the ddc module format. This will replace all other
existing formats long-term. For now, we want to enable some test
bots to start running with the new format, so we emit different
HTML based on the format in the test runner.

With these changes, all tests pass when using ddc module format
except for libraries_test. This is due to how DDC may lazily load
libraries with the new format, so this test is reordered to make
it work with the new format.

Change-Id: If52bb64977f96b4c4a287f142026412a4db558cd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/379750
Reviewed-by: Bob Nystrom <rnystrom@google.com>
Commit-Queue: Srujan Gaddam <srujzs@google.com>
Reviewed-by: Nicholas Shahan <nshahan@google.com>