[dart2wasm] Use static dispatch on common, very fast OneByteString operations
Most strings are [OneByteString]s and operations such as

  * length
  * isEmpty
  * isNonEmpty
  * charCodeAt

are so quick that dispatch table overhead is enourmous compared to the
actual logic.

=> Use @pragma('wasm:static-dispatch') to avoid dispatch table overhead
=> Use @pragma('wasm:prefer-inline') to get them inlined

(This is a workaround the fact that V8 doesn't have speculative
inlining yet)

This makes certain json decoding benchmarks XX % faster.

Change-Id: Ie639343f3ae1f3ce9bd2f20d2d1b8c2bd88a52d8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/380821
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>