[dart2wasm] Use Wasm arrays directly for `_JsonListener` state
Improves bytes/second in an internal JSON decoding benchmark from
227,791,937 to 246,298,043 (+8%).

The container stack (`_JsonListener.stack`) could hold the pair `(array,
array length)`, however since records are also heap allocated, and size
of the record would be the same as `GrowableList`, I've reused the
`GrowableList` type for the stack elements.

To avoid duplicating the code for growing the Wasm arrays in class
fields, a new intrinsic `pushWasmArray` implemented as a macro. It's
implemented as a kernel transform and it copies the arguments as
expressions in the transformed code to grow the Wasm array when it's
full.

To change as little is possible in one CL (to be able to measure
impact), growth strategy for the arrays are kept the same.

Change-Id: I2e6ba7704a83e6e6d8ada5d9015e208069737170
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/380705
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>