[dart2wasm] Move Wasm compilation functions into generated runtime blob
To be able to control how dart2wasm-generated modules are compiled, add
a `WebAssembly.compile` and `WebAssmebly.compileStreaming` wrapper to
the generated runtime blob.

These wrappers do js-string built-in detection and enable the built-in
when it's available.

Once the SDK is rolled into engine we will update the engine dart2wasm
module loader to use `compileStreaming` from the runtime blob, in
https://github.com/flutter/engine/pull/51488.

This will allow enabling the built-in without duplicating the feature
detection and compilation code in the engine.

Change-Id: Icb939f16a2b0ae1d348ec5a814371e9ffde4fd68
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/381581
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>