[dart2wasm] Fix instantiation context types in instantiation trampolines
Type for a closure with one argument (Closure-0-1) is subtype of type of
a closure with no arguments (Closure-0-0).

This allows calling the closure with no arguments in the generated code.

(This only happens when the Dart types align, i.e. the argument is
optional)

However instantiation vtable functions for e.g. Closure-1-1 currently
can't be as Closure-1-0, resulting in a type cast error in a code like:

    void test<T>([T? x]) {}

    void Function() x = runtimeTrue ? test : () {};
    x();

Relax the instantiation context struct type in instantiation vtable
functions and downcast the generic closure struct type (type of the
instantiation closure) to the right type before calling instantiation
clsoure.

Fixes #56534.

Change-Id: I59f9b019c4dc8d2e26ee21f1f6c3c607af0be4d8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/381641
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>