[dart2wasm] Remove makeListFixedLength/makeFixedListUnmodifiable
This removes usages of makeListFixedLength/makeFixedListUnmodifiable
in wasm core library implementation.

Furthermore we establish a guarantee that the backing store of
fixed length lists (modifiable or unmodifiable) always has the same
size as the actual list (and not larger).

This in return allows the fixed length iterators to not need a separate
length to track, as the backing WasmArray<Object?>'s length is
sufficient.

Change-Id: I1fe44781c6044187a3a129615a5065f08dfc53a7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/382860
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>